split up common.cpp into separate cpp and hpp files; will re-combine as necessary. re-wrote ~90% of main.cpp using more efficient and original code. Bare bones rendering (no textures) and cube generation.
