+++ Makefile
@@ -9,19 +9,18 @@ SRC_DIR := src
 INCLUDES = -Isrc/include -I/usr/include/freetype2
 LINKER_FLAGS := -lglfw
 
-# SRCS := $(shell find $(SRC_DIR) -name '*.cpp' -or -name '*.c')
 SRCS := src/main.cpp src/glad.c
 
-# test:
-# 	 g++ -lglfw -Isrc/include ./src/temp_main.cpp ./src/glad.c -o test
-# 	./test
-# 	rm ./test
+# FPS limit for custom mangohud test run
+FPS_LIMIT := 60
+
 
 test:
 	$(CXX) $(LINKER_FLAGS) $(INCLUDES) $(SRCS) -o $(TEST_NAME)
 
 testrun: test
-	./$(TEST_NAME)
+# 	./$(TEST_NAME) # Normal run without mangohud
+	~/bin/mangohudtest $(FPS_LIMIT) ./$(TEST_NAME) # Run using custom mangohud command to limit FPS for test cases
 	rm ./$(TEST_NAME)
 
 .PHONY: build test testrun clean
