--- src/common.cpp
@@ -1,188 +1,0 @@
-#include <fstream>
-#include <sstream>
-#include <iostream>
-
-#include <common.hpp>
-
-#define STB_IMAGE_IMPLEMENTATION
-#include <stb_image.h>
-
-void shaderErrorHandler(unsigned int shader, bool is_program)
-{
-	int success;
-	char info_log[512];
-
-	if(is_program)
-	{
-		glGetProgramiv(shader, GL_LINK_STATUS, &success);
-		if(!success) {
-			glGetProgramInfoLog(shader, 512, NULL, info_log);
-			std::cout << "ERROR::SHADER::PROGRAM::LINKING_FAILED\n" << info_log << std::endl;
-		}
-	}
-	else
-	{
-		glGetShaderiv(shader, GL_COMPILE_STATUS, &success);
-		if(!success)
-		{
-			glGetShaderInfoLog(shader, 512, NULL, info_log);
-			std::cout << "ERROR::SHADER::VERTEX::COMPILATION_FAILED\n" << info_log << std::endl;
-		}
-	}
-}
-
-//
-// Shader Class
-//
-
-Shader::Shader(const char* vertex_path, const char* fragment_path)
-{
-	std::string vertex_code;
-	std::string fragment_code;
-	std::ifstream v_shader_file;
-	std::ifstream f_shader_file;
-
-	v_shader_file.exceptions(std::ifstream::failbit | std::ifstream::badbit);
-	f_shader_file.exceptions(std::ifstream::failbit | std::ifstream::badbit);
-	try
-	{
-		v_shader_file.open(vertex_path);
-		f_shader_file.open(fragment_path);
-		std::stringstream v_shader_stream, f_shader_stream;
-
-		v_shader_stream << v_shader_file.rdbuf();
-		f_shader_stream << f_shader_file.rdbuf();
-
-		v_shader_file.close();
-		f_shader_file.close();
-
-		vertex_code = v_shader_stream.str();
-		fragment_code = f_shader_stream.str();
-	}
-
-	catch(std::ifstream::failure e)
-	{
-		std::cout << "ERROR::SHADER::FILE_NOT_SUCCESFULLY_READ" << std::endl;
-	}
-
-	const char* v_shader_code = vertex_code.c_str();
-	const char* f_shader_code = fragment_code.c_str();
-
-	unsigned int vertex, fragment;
-	vertex = glCreateShader(GL_VERTEX_SHADER);
-	glShaderSource(vertex, 1, &v_shader_code, NULL);
-	glCompileShader(vertex);
-	shaderErrorHandler(vertex);
-
-	fragment = glCreateShader(GL_FRAGMENT_SHADER);
-	glShaderSource(fragment, 1, &f_shader_code, NULL);
-	glCompileShader(fragment);
-	shaderErrorHandler(fragment);
-
-	ID = glCreateProgram();
-	glAttachShader(ID, vertex);
-	glAttachShader(ID, fragment);
-	glLinkProgram(ID);
-	shaderErrorHandler(ID, true);
-
-	glDeleteShader(vertex);
-	glDeleteShader(fragment);
-};
-
-void Shader::use()
-{
-	glUseProgram(ID);
-}
-
-void Shader::setBool(const std::string &name, bool value) const
-{
-	glUniform1i(glGetUniformLocation(ID, name.c_str()), (int)value);
-}
-
-void Shader::setInt(const std::string &name, int value) const
-{
-	glUniform1i(glGetUniformLocation(ID, name.c_str()), value);
-}
-
-void Shader::setFloat(const std::string &name, float value) const
-{
-	glUniform1f(glGetUniformLocation(ID, name.c_str()), value);
-}
-
-//
-// Camera Class
-//
-
-Camera::Camera(glm::vec3 position, glm::vec3 up, float yaw, float pitch) : Front(glm::vec3(0.0f, 0.0f, -1.0f)), MovementSpeed(SPEED), MouseSensitivity(SENSITIVITY)
-{
-	Position = position;
-	WorldUp = up;
-	Yaw = yaw;
-	Pitch = pitch;
-	updateCameraVectors();
-}
-
-Camera::Camera(float posX, float posY, float posZ, float upX, float upY, float upZ, float yaw, float pitch) : Front(glm::vec3(0.0f, 0.0f, -1.0f)), MovementSpeed(SPEED), MouseSensitivity(SENSITIVITY)
-{
-	Position = glm::vec3(posX, posY, posZ);
-	WorldUp = glm::vec3(upX, upY, upZ);
-	Yaw = yaw;
-	Pitch = pitch;
-	updateCameraVectors();
-}
-
-glm::mat4 Camera::GetViewMatrix()
-{
-	return glm::lookAt(Position, Position + Front, Up);
-}
-
-void Camera::ProcessKeyboard(Camera_Movement direction, float deltaTime)
-{
-	float velocity = MovementSpeed * deltaTime;
-	switch (direction)
-	{
-		case FORWARD:
-			Position += Front * velocity;
-			break;
-		case BACKWARD:
-			Position -= Front * velocity;
-			break;
-		case LEFT:
-			Position -= Right * velocity;
-			break;
-		case RIGHT:
-			Position += Right * velocity;
-			break;
-	}
-}
-
-void Camera::ProcessMouseMovement(float xoffset, float yoffset, GLboolean constrainPitch = true)
-{
-	xoffset *= MouseSensitivity;
-	yoffset *= MouseSensitivity;
-
-	Yaw += xoffset;
-	Pitch += yoffset;
-
-	if (constrainPitch)
-	{
-		if (Pitch > 89.0f)
-			Pitch = 89.0f;
-		if (Pitch < -89.0f)
-			Pitch = -89.0f;
-	}
-
-	updateCameraVectors();
-}
-
-void Camera::updateCameraVectors()
-{
-	glm::vec3 front;
-	front.x = cos(glm::radians(Yaw)) * cos(glm::radians(Pitch));
-	front.y = sin(glm::radians(Pitch));
-	front.z = sin(glm::radians(Yaw)) * cos(glm::radians(Pitch));
-	Front = glm::normalize(front);
-
-	Right = glm::normalize(glm::cross(Front, WorldUp));
-	Up = glm::normalize(glm::cross(Right, Front));
-}
