+++ src/include/g_common.cpp
@@ -1,0 +1,1 @@
+#include <g_common.hpp>
