+++ src/include/g_common.hpp
@@ -1,0 +1,18 @@
+#ifndef COMMON_INCLUDES
+#define COMMON_INCLUDES
+#include <iostream>
+#include <fstream>
+#include <sstream>
+#include <string>
+#include <map>
+#endif
+
+#ifndef GL_INCLUDES
+#define GL_INCLUDES
+#include <glad/glad.h>
+#include <glm/fwd.hpp>
+#include <GLFW/glfw3.h>
+#include <glm/glm.hpp>
+#include <glm/gtc/matrix_transform.hpp>
+#include <glm/gtc/type_ptr.hpp>
+#endif
