+++ src/include/g_cube.hpp
@@ -1,0 +1,48 @@
+#include <g_common.hpp>
+
+#ifndef CUBE_CLASS_DEFINITION
+#define CUBE_CLASS_DEFINITION
+class Cube
+{
+public:
+	float vertices[24] = {
+		-1.0f, -1.0f, -1.0f,	// Left		Down	Back		0
+		-1.0f, -1.0f, 1.0f,		// Left		Down	Forward		1
+		-1.0f, 1.0f, -1.0f,		// Left		Up		Back		2
+		-1.0f, 1.0f, 1.0f,		// Left		Up		Forward		3
+		 1.0f, -1.0f, -1.0f,	// Right	Down	Back		4
+		 1.0f, -1.0f, 1.0f,		// Right	Down	Forward		5
+		 1.0f, 1.0f, -1.0f,		// Right	Up		Back		6
+		 1.0f, 1.0f, 1.0f,		// Right	Up		Forward		7
+	};
+
+	unsigned int indices[36] = {
+		// -X Square
+		0, 1, 2,	// Bottom Triangle
+		3, 1, 2,	// Top Triangle
+
+		// +X Square
+		4, 5, 6,
+		7, 5, 6,
+
+		// -Y Square
+		0, 1, 4,
+		5, 1, 4,
+
+		// +Y Square
+		2, 3, 6,
+		7, 3, 6,
+
+		// -Z Square
+		0, 2, 4,
+		6, 2, 4,
+
+		// +Z Square
+		1, 3, 5,
+		7, 3, 5
+	};
+
+	void makeCube();
+	void drawCube();
+};
+#endif
