+++ src/include/g_shader.cpp
@@ -1,0 +1,109 @@
+#include <g_shader.hpp>
+#include <glm/gtc/type_ptr.hpp>
+
+//
+// Shader Class
+//
+
+Shader::Shader(const char* vertex_path, const char* fragment_path)
+{
+	std::string vertex_code;
+	std::string fragment_code;
+	std::ifstream v_shader_file;
+	std::ifstream f_shader_file;
+
+	v_shader_file.exceptions(std::ifstream::failbit | std::ifstream::badbit);
+	f_shader_file.exceptions(std::ifstream::failbit | std::ifstream::badbit);
+	try
+	{
+		v_shader_file.open(vertex_path);
+		f_shader_file.open(fragment_path);
+		std::stringstream v_shader_stream, f_shader_stream;
+
+		v_shader_stream << v_shader_file.rdbuf();
+		f_shader_stream << f_shader_file.rdbuf();
+
+		v_shader_file.close();
+		f_shader_file.close();
+
+		vertex_code = v_shader_stream.str();
+		fragment_code = f_shader_stream.str();
+	}
+
+	catch(std::ifstream::failure e)
+	{
+		std::cout << "ERROR::SHADER::FILE_NOT_SUCCESFULLY_READ" << std::endl;
+	}
+
+	const char* v_shader_code = vertex_code.c_str();
+	const char* f_shader_code = fragment_code.c_str();
+
+	unsigned int vertex, fragment;
+	vertex = glCreateShader(GL_VERTEX_SHADER);
+	glShaderSource(vertex, 1, &v_shader_code, NULL);
+	glCompileShader(vertex);
+	shaderErrorHandler(vertex);
+
+	fragment = glCreateShader(GL_FRAGMENT_SHADER);
+	glShaderSource(fragment, 1, &f_shader_code, NULL);
+	glCompileShader(fragment);
+	shaderErrorHandler(fragment);
+
+	ID = glCreateProgram();
+	glAttachShader(ID, vertex);
+	glAttachShader(ID, fragment);
+	glLinkProgram(ID);
+	shaderErrorHandler(ID, true);
+
+	glDeleteShader(vertex);
+	glDeleteShader(fragment);
+};
+
+void Shader::use()
+{
+	glUseProgram(ID);
+}
+
+void Shader::setBool(const std::string &name, bool value) const
+{
+	glUniform1i(glGetUniformLocation(ID, name.c_str()), (int)value);
+}
+
+void Shader::setInt(const std::string &name, int value) const
+{
+	glUniform1i(glGetUniformLocation(ID, name.c_str()), value);
+}
+
+void Shader::setFloat(const std::string &name, float value) const
+{
+	glUniform1f(glGetUniformLocation(ID, name.c_str()), value);
+}
+
+void Shader::setMatrix(const std::string &name, glm::mat4 value) const
+{
+	glUniformMatrix4fv(glGetUniformLocation(ID, name.c_str()), 1, GL_FALSE, glm::value_ptr(value));
+}
+
+void Shader::shaderErrorHandler(unsigned int shader, bool is_program)
+{
+	int success;
+	char info_log[512];
+
+	if(is_program)
+	{
+		glGetProgramiv(shader, GL_LINK_STATUS, &success);
+		if(!success) {
+			glGetProgramInfoLog(shader, 512, NULL, info_log);
+			std::cout << "ERROR::SHADER::PROGRAM::LINKING_FAILED\n" << info_log << std::endl;
+		}
+	}
+	else
+	{
+		glGetShaderiv(shader, GL_COMPILE_STATUS, &success);
+		if(!success)
+		{
+			glGetShaderInfoLog(shader, 512, NULL, info_log);
+			std::cout << "ERROR::SHADER::VERTEX::COMPILATION_FAILED\n" << info_log << std::endl;
+		}
+	}
+}
