--- src/include/mathlib.hpp
@@ -1,6 +1,0 @@
-#ifndef FUNCTIONS_MATH_INCLUDED
-#define FUNCTIONS_MATH_INCLUDED
-
-float J_rsqrt(float number);
-
-#endif
