+++ src/main.cpp
@@ -1,18 +1,17 @@
-#include <iostream>
-#include <fstream>
-#include <sstream>
+#include <g_common.hpp>
+#include <g_cube.cpp>
+#include <g_shader.cpp>
+#include <g_camera.cpp>
 
-#include <glad/glad.h>
-#include <glm/fwd.hpp>
-#include <GLFW/glfw3.h>
-#include <glm/glm.hpp>
-#include <glm/gtc/matrix_transform.hpp>
-#include <glm/gtc/type_ptr.hpp>
+Camera player_camera(glm::vec3(0.0f, 0.0f, 3.0f));
 
 unsigned int window_width = 1280, window_height = 720;
+float mouse_last_x = (int)(window_width / 2), mouse_last_y = (int)(window_height / 2);
+float delta_time = 0.0f, last_frame = 0.0f;
 
-void framebufferSizeCallback(GLFWwindow* window, int width, int height);
 void processInput(GLFWwindow* window);
+void framebufferSizeCallback(GLFWwindow* window, int width, int height);
+void mouseCallback(GLFWwindow* window, double x_position, double y_position);
 
 int main(int argc, char** argv)
 {
@@ -32,8 +31,9 @@ int main(int argc, char** argv)
 
 	glfwMakeContextCurrent(window);
 	glfwSetWindowPos(window, (int)((1920 - window_width) / 2), (int)((1080 - window_height) /2));
-	// glfwSetInputMode(window, GLFW_CURSOR, GLFW_CURSOR_DISABLED);
 	glfwSetFramebufferSizeCallback(window, framebufferSizeCallback);
+	glfwSetInputMode(window, GLFW_CURSOR, GLFW_CURSOR_DISABLED);
+	glfwSetCursorPosCallback(window, mouseCallback);
 
 	if (!gladLoadGLLoader((GLADloadproc)glfwGetProcAddress))
 	{
@@ -41,18 +41,57 @@ int main(int argc, char** argv)
 		return -1;
 	}
 
-	glEnable(GL_DEPTH_TEST);
+	Shader simple_shader("src/shaders/vertex_shader.glsl", "src/shaders/fragment_shader.glsl");
+	Cube new_cube;
 
+	unsigned int element_buffer, vertex_array, vertex_buffer;
+	glGenBuffers(1, &element_buffer);
+	glGenVertexArrays(1, &vertex_array);
+	glGenBuffers(1, &vertex_buffer);
+
+	glBindVertexArray(vertex_array);
+
+	glBindBuffer(GL_ARRAY_BUFFER, vertex_buffer);
+	glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, element_buffer);
+
+	new_cube.makeCube();
+
+	glEnable(GL_DEPTH_TEST);
 	// glPolygonMode(GL_FRONT_AND_BACK, GL_LINE);	// Wireframe mode
+
+	//
+	// Runtime Loop
+	//
 	while(!glfwWindowShouldClose(window))
 	{
+		float current_frame = glfwGetTime();
+		delta_time = current_frame - last_frame;
+		last_frame = current_frame;
+
 		processInput(window);
 		glClearColor(0.2f, 0.2f, 0.5f, 1.0f);
 		glClear(GL_COLOR_BUFFER_BIT | GL_DEPTH_BUFFER_BIT);
+
+		glm::mat4 model_location = glm::mat4(1.0f);
+		glm::mat4 projection = glm::perspective(glm::radians(45.0f), (float)window_width / (float)window_height, 0.1f, 100.0f);
+		glm::mat4 camera_view = player_camera.GetViewMatrix();
+
+		simple_shader.setMatrix("model", model_location);
+		simple_shader.setMatrix("projection", projection);
+		simple_shader.setMatrix("camera_view", camera_view);
+		simple_shader.use();
+
+		glBindVertexArray(vertex_array);
+		new_cube.drawCube();
+
 		glfwSwapBuffers(window);
 		glfwPollEvents();
 	}
 
+	glDeleteVertexArrays(1, &vertex_array);
+	glDeleteBuffers(1, &vertex_buffer);
+	glDeleteBuffers(1, &element_buffer);
+	glDeleteProgram(simple_shader.ID);
 	glfwTerminate();
 	return 0;
 }
@@ -68,4 +107,27 @@ void processInput(GLFWwindow* window)
 {
 	if (glfwGetKey(window, GLFW_KEY_ESCAPE) == GLFW_PRESS)
 		glfwSetWindowShouldClose(window, true);
+
+	if (glfwGetKey(window, GLFW_KEY_W) == GLFW_PRESS)
+		player_camera.ProcessKeyboard(FORWARD, delta_time);
+	if (glfwGetKey(window, GLFW_KEY_S) == GLFW_PRESS)
+		player_camera.ProcessKeyboard(BACKWARD, delta_time);
+	if (glfwGetKey(window, GLFW_KEY_D) == GLFW_PRESS)
+		player_camera.ProcessKeyboard(RIGHT, delta_time);
+	if (glfwGetKey(window, GLFW_KEY_A) == GLFW_PRESS)
+		player_camera.ProcessKeyboard(LEFT, delta_time);
+}
+
+void mouseCallback(GLFWwindow* window, double x_position_in, double y_position_in)
+{
+
+	float x_position = static_cast<float>(x_position_in);
+	float y_position = static_cast<float>(y_position_in);
+
+	float x_offset = x_position - mouse_last_x;
+	float y_offset = mouse_last_y - y_position;
+	mouse_last_x = x_position;
+	mouse_last_y = y_position;
+
+	player_camera.ProcessMouseMovement(x_offset, y_offset);
 }
