+++ src/shaders/fragment_shader.glsl
@@ -1,14 +1,11 @@
 #version 460 core
 out vec4 FragColor;
 
-// in vec3 vertexColor;
-in vec2 textureCoordinates;
+// in vec2 textureCoordinates;
 
-uniform sampler2D newTexture2D;
+// uniform sampler2D newTexture2D;
 
 void main()
 {
-	// FragColor = vec4(vertexColor, 1.0);
-	// FragColor = texture(newTexture2D, textureCoordinates) * vec4(vertexColor, 1.0);
-	FragColor = texture(newTexture2D, textureCoordinates);
+	FragColor = vec4(0.0, 1.0, 0.0, 1.0);
 };
